requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.1
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
